Move realtime viewport code into separate experimental file
